+++ met-web/src/components/Form/FormSubmit.tsx
@@ -1,39 +1,15 @@
-/* eslint-disable */
-import React, { useState } from 'react';
-import { Form } from '@formio/react';
-import { FormSubmissionData, FormSubmitterProps } from './types';
-import FormStepper from 'components/survey/submit/Stepper';
-interface PageData {
-    page: number;
-    submission: unknown;
-}
+import React from 'react';
+import { FormSubmitterProps } from './types';
+import SinglePageForm from './SinglePageForm';
+import MultiPageForm from './MultiPageForm';
 
 const FormSubmit = ({ handleFormChange, savedForm, handleFormSubmit }: FormSubmitterProps) => {
-    const [currentPage, setCurrentPage] = useState(0);
     const isMultiPage = savedForm && savedForm.display === 'wizard';
 
-    return (
-        <div className="formio">
-            {isMultiPage ? (
-                <FormStepper
-                    currentPage={currentPage}
-                    totalPages={savedForm?.components.length - 1}
-                    pages={savedForm?.components}
-                />
-            ) : (
-                <></>
-            )}
-            <Form
-                form={savedForm || { display: 'form' }}
-                onChange={(form: unknown) => handleFormChange(form as FormSubmissionData)}
-                onNextPage={(pageData: PageData) => setCurrentPage(pageData.page)}
-                onPrevPage={(pageData: PageData) => setCurrentPage(pageData.page)}
-                onSubmit={(form: unknown) => {
-                    const formSubmissionData = form as FormSubmissionData;
-                    handleFormSubmit(formSubmissionData.data);
-                }}
-            />
-        </div>
+    return isMultiPage ? (
+        <MultiPageForm handleFormChange={handleFormChange} savedForm={savedForm} handleFormSubmit={handleFormSubmit} />
+    ) : (
+        <SinglePageForm handleFormChange={handleFormChange} savedForm={savedForm} handleFormSubmit={handleFormSubmit} />
     );
 };
