+++ met-web/tests/unit/components/engagement/EngagementForm.test.tsx
@@ -73,9 +73,6 @@ describe('Engagement form page tests', () => {
     const getEngagementMetadataMock = jest
         .spyOn(engagementMetadataService, 'getEngagementMetadata')
         .mockReturnValue(Promise.resolve(engagementMetadata));
-    const patchEngagementMetadataMock = jest
-        .spyOn(engagementMetadataService, 'patchEngagementMetadata')
-        .mockReturnValue(Promise.resolve(engagementMetadata));
     const getEngagementMock = jest
         .spyOn(engagementService, 'getEngagement')
         .mockReturnValue(Promise.resolve(draftEngagement));
